// src/entities/puzzle/puzzle.enum.ts
export enum PuzzleEditorRank {
  Iron = 'Iron',
  Bronze = 'Bronze',
  Silver = 'Silver',
  Gold = 'Gold',
  Platinum = 'Platinum',
  Emerald = 'Emerald',
  Diamond = 'Diamond',
  Master = 'Master',
  GrandMaster = 'GrandMaster',
  Challenger = 'Challenger',
  UNRANKED = 'unranked',
}

export enum RiotRank {
  Iron = 'IRON',
  Bronze = 'BRONZE',
  Silver = 'SILVER',
  Gold = 'GOLD',
  Platinum = 'PLATINUM',
  Emerald = 'EMERALD',
  Diamond = 'DIAMOND',
  Master = 'MASTER',
  GrandMaster = 'GRANDMASTER',
  Challenger = 'CHALLENGER',
  UNRANKED = 'Unranked',
}

export enum PuzzleEditorRole {
  Sup = 'Support',
  Adc = 'Adc',
  Mid = 'Mid',
  Jngl = 'Jungle',
  Top = 'Top',
}

export enum RiotRole {
  Top = 'TOP',
  Mid = 'MIDDLE',
  Bot = 'BOTTOM',
  Jungle = 'JUNGLE',
  Support = 'UTILITY',
}

export enum Difficulty {
  Easy = 'Easy',
  Medium = 'Medium',
  Hard = 'Hard',
}

export enum PuzzleState {
  UNPUBLISHED = 'unpublished',
  PUBLISHED = 'published',
  PROCESSING = 'processing',
  ARCHIVED = 'archived',
}

export enum Champion {
  ASHE = 'ashe',
  AZIR = 'azir',
  CORKI = 'corki',
}